+++ hazelcast5-consumer/src/main/java/no/mnemonic/services/common/hazelcast/consumer/KafkaToHazelcastHandler.java
@@ -1,14 +1,18 @@
 package no.mnemonic.services.common.hazelcast.consumer;
 
 import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.core.TransactionalQueue;
 import com.hazelcast.transaction.TransactionContext;
 import com.hazelcast.transaction.TransactionOptions;
+import com.hazelcast.transaction.TransactionalQueue;
 import no.mnemonic.commons.component.Dependency;
 import no.mnemonic.commons.component.LifecycleAspect;
 import no.mnemonic.commons.logging.Logger;
 import no.mnemonic.commons.logging.Logging;
-import no.mnemonic.commons.metrics.*;
+import no.mnemonic.commons.metrics.MetricAspect;
+import no.mnemonic.commons.metrics.MetricException;
+import no.mnemonic.commons.metrics.Metrics;
+import no.mnemonic.commons.metrics.MetricsData;
+import no.mnemonic.commons.metrics.TimerContext;
 import no.mnemonic.commons.utilities.collections.CollectionUtils;
 import no.mnemonic.messaging.documentchannel.DocumentBatch;
 import no.mnemonic.messaging.documentchannel.DocumentSource;
@@ -34,10 +38,7 @@ import static no.mnemonic.commons.utilities.lambda.LambdaUtils.tryTo;
  * which means the offset is moved forward in Kafka.
  * Hazelcast queue is used to store the messages from Kafka to guarantee that the messages won't be lost due to e.g. current server goes down,
  * so that the Hazelcast queue consumer can poll from it.
- *
- * @deprecated Use <code>hazelcast5-consumer</code> package instead
  */
-@Deprecated
 public class KafkaToHazelcastHandler<T> implements LifecycleAspect, MetricAspect {
 
   private static final Logger LOGGER = Logging.getLogger(KafkaToHazelcastHandler.class);