+++ hazelcast5-consumer/src/main/java/no/mnemonic/services/common/hazelcast/consumer/TransactionalConsumer.java
@@ -6,10 +6,6 @@ import java.io.Closeable;
 import java.io.IOException;
 import java.util.Collection;
 
-/**
- * @deprecated Use <code>hazelcast5-consumer</code> package instead
- */
-@Deprecated
 public interface TransactionalConsumer<T> extends Closeable {
 
   void consume(Collection<T> items) throws IOException, InterruptedException, ConsumerGaveUpException;